restructured project. split bigger 500+ line files based on related functionality.
