--- cmd/web/routes.go
@@ -1,10 +1,0 @@
-package main
-
-// func routes() *http.ServeMux {
-// 	mux := http.NewServeMux()
-// 	fileServer := http.FileServer(http.Dir("./ui/static/"))
-// 	mux.Handle("/static/", http.StripPrefix("/static", fileServer))
-// 	mux.HandleFunc("/", home)
-// 	//	mux.HandleFunc("/task/create", taskCreate)
-// 	return mux
-// }
